pyproject.toml
[tool.poetry]
name = "Python.Study"
version = "0.1.36"
description = ""
authors = ["解大劦 <develope@163.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
basiclibrary-py = "^0.6.13"
tomlkit = "^0.12.4"
pyautogui = "^0.9.54"
xpinyin = "^0.7.6"
superstream = "^0.2.6"
streamlit = "^1.35.0"
ulid-py = "^1.1.0"
pypattyrn = "^1.2"

[tool.poetry.group.dev.dependencies]
pylint = "^3.1.0"
pytest = "^8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# 使用国内的类库镜像源
[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"
